pyproject.toml
[tool.poetry]
name = "ssu-hackaton-backend2023"
version = "0.1.0"
description = ""
authors = ["Nikita Rydanov <wrongwayboyy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
docker-compose = "^1.29.2"
sqlalchemy = "^2.0.11"
flask = "^2.3.1"
greenlet = "^2.0.2"
psycopg2 = "^2.9.6"
flask-cors = "^3.0.10"
pyjwt = "^2.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
